// types/buildings.ts
export interface Building {
  id: number
  developer_id: number
  name: string
  location: string
  micromarket_zone: string
  building_structure: string
  building_title: string
  grade: string
  total_area: number
  certifications: string
  google_coordinates: string
  cam: number
  year_built: number
  construction_status: string
  building_status: string
  building_image_link: string
}